pyproject.toml
[project]
name = "book-store"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "streamlit (>=1.42.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
